requirements.txt
absl-py==1.4.0
anyio==3.6.2
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
dnspython==2.3.0
email-validator==1.3.1
executing==1.2.0
fastapi==0.92.0
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mpmath==1.2.1
numpy==1.24.2
oauthlib==3.2.2
onnx==1.13.1
onnxruntime==1.14.0
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
orjson==3.8.6
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.37
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.5
Pygments==2.14.0
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.15.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.25.0
sympy==1.11.1
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
# tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
thop==0.1.1.post2209072238
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.7.0
ultralytics==8.0.20
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.14.1
zipp==3.14.0
